+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+### Modified
+- rails gems update
+
+### Fixed
+- bug with guild invites to events for not main characters
+
 ## [1.3.4] - 2020-05-02
 ### Modified
 - for line up show only approved characters in counter